% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_soccerpitch.R
\name{draw_soccerpitch}
\alias{draw_soccerpitch}
\title{Ggplot2 Soccer Field Markings}
\usage{
draw_soccerpitch(xmin = 0, xmax = 100 * 2/3, ymin = 0, ymax = 100,
  thickness = 1.5, markings = "full", palette = "classic",
  layout = "landscape", portion = "full")
}
\arguments{
\item{xmin}{The origin of x for the markings to begin. X is 'across' i.e. along the halfway line}

\item{xmax}{The maximum value of x for the markings to end}

\item{ymin}{The origin of y for the markings to begin. Y is 'along' i.e. up the touchline}

\item{ymax}{The maximum value of y for the markings to end}

\item{thickness}{The thickness of the markings}

\item{markings}{Which markings to plot from the most basic rectangle to full pitch markings}

\item{palette}{The palette to use to colour the field/markings.
Standard arguments are classic (the 'natural' colours), bw (a greyscale plot), or blues (darkblue),
but custom palettes can be supplied}

\item{layout}{The orientation of the final plot. Defaults to landscape (so coord_flip() s the plot)}

\item{portion}{The portion of the field to plot. 
Standard arguments are the whole pitch (full) or half-pitch (half) but custom coordinates can be supplied.}
}
\description{
Ggplot2 Soccer Field Markings
}
\author{
Robert Hickman
}
