% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sportsfield.R
\name{theme_sportsfield}
\alias{theme_sportsfield}
\title{Sportsfield Ggplot2 Theme}
\usage{
theme_sportsfield(ratio = 1)
}
\arguments{
\item{ratio}{The aspect ration of the plot to output}
}
\description{
Sportsfield Ggplot2 Theme
}
\author{
Robert Hickman
}
